Moved service decorator to service helpers
Moved the service decorator to the service helpers module and moved the
associated tests.
